requirements.txt
requests==2.33.0
pydantic==2.12.5
pyyaml==6.0.3